% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{dataset_loader}
\alias{dataset_loader}
\title{Create data sets and dataloaders}
\usage{
dataset_loader(train_x, train_y, test_x, test_y, batch_size = 50, dev)
}
\arguments{
\item{train_x}{numeric, scaled matrix of predictors used for training}

\item{train_y}{numeric, scaled vector of target used for training}

\item{test_x}{integer, matrix of predictors used for testing}

\item{test_y}{integer, vector of predictors used for testing}

\item{batch_size}{integer indicating a batch size used in dataloader. Default: 50}

\item{dev}{device used to calculations (cpu or gpu)}
}
\value{
list of two data sets and two dataloaders for train and test respectively
}
\description{
Create two data sets, from given ..x matrices and ..y vectors and converts them into
dataloaders with provided batch size.
}
