% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairtrain.R
\name{fairtrain}
\alias{fairtrain}
\title{Proceeds with fair training with PAN model}
\usage{
fairtrain(
  train_x,
  test_x,
  train_y,
  test_y,
  sensitive_train,
  sensitive_test,
  BATCH_SIZE = 50,
  NEURONS_CLF = c(32, 32, 32),
  NEURONS_ADV = c(32, 32, 32),
  DIMENSION_CLF = 2,
  DMIENSION_ADV = 2,
  N_EP_PRECLF = 5,
  LEARNING_RATE_CLF = 0.001,
  PROTECTED,
  PRIVILIGED,
  PARTITION = 0.7,
  N_EP_PAN = 50,
  LEARNING_RATE_ADV = 0.001,
  LAMBDA = 50,
  DATA,
  N_EP_PREADV = 10
)
}
\arguments{
\item{train_x}{numeric, scaled matrix of predictors used for training}

\item{test_x}{integer, matrix of predictors used for testing}

\item{train_y}{numeric, scaled vector of target used for training}

\item{test_y}{integer, vector of predictors used for testing}

\item{sensitive_train}{integer, vector of sensitive values used for training}

\item{sensitive_test}{integer, vector of sensitive values used for testing}

\item{BATCH_SIZE}{integer indicating a batch size used in dataloader. Default: 50}

\item{NEURONS_CLF}{integer vector describing a neural architecture of classifier
network. Default: c(32,32,32). This notation means that the network
has 3 layers with 32 neurons each.}

\item{NEURONS_ADV}{integer vector describing a neural architecture of adversarial
network. Default: c(32,32,32). This notation means that the network
has 3 layers with 32 neurons each.}

\item{DIMENSION_CLF}{integer from [0,2] setting nnf_softmax dimension for classifier.
Default: 2}

\item{DMIENSION_ADV}{integer from [0,2] setting nnf_softmax dimension for adversarial.
Default: 2}

\item{N_EP_PRECLF}{integer setting number of epochs for preclassifiers training.
Default: 5}

\item{LEARNING_RATE_CLF}{float from [0,1] setting learning rate for classifier.
Default: 0.001}

\item{PROTECTED}{numerical vector of sensitive variables}

\item{PRIVILIGED}{string label of privileged class in protected}

\item{PARTITION}{float from [0,1] range setting the size of train vector (test size
equals 1-PARTITION). Default = 0.7.}

\item{N_EP_PAN}{integer setting number of epochs for PAN training. Default : 50}

\item{LEARNING_RATE_ADV}{float from [0,1] setting learning rate for classifier.
Default: 0.001}

\item{LAMBDA}{integer parameter regulating learning proccess (intuition: the bigger it is,
the fairer predictions). Default: 50}

\item{DATA}{numerical table of predictors}

\item{N_EP_PREADV}{integer setting number of epochs for preadversarials training.
Default : 10}
}
\value{

}
\description{
Core function of the whole package. It trains a PAN model with fair predictions.
}
