% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{make_preds}
\alias{make_preds}
\title{Makes binary/probabilistic predictions}
\usage{
make_preds(model, test_ds)
}
\arguments{
\item{model}{neural network classification model}

\item{test_ds}{dataset object used for making test predictions}
}
\value{
integer(binary) or float(probabilistic) vector of predictions
}
\description{
These two functions make binary or probabilistic predctions of provided model based on
test data set
}
