% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{make_preds_prob}
\alias{make_preds_prob}
\title{Makes probabilistic predictions}
\usage{
make_preds_prob(model, test_ds, dev)
}
\arguments{
\item{model}{neural network classification model}

\item{test_ds}{dataset object used for making test predictions}

\item{dev}{device used to calculations (cpu or gpu)}
}
\value{
float (probabilistic) vector of predictions
}
\description{
These two functions make probabilistic predictions of provided model based on test
data set.
}
