% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_that.R
\name{plot_fairness}
\alias{plot_fairness}
\title{Plots fobject metrics}
\usage{
plot_fairness(fobject)
}
\arguments{
\item{fobject}{fairness object from fairmodels}
}
\value{
NULL - plots the visualization
}
\description{
This function plots fairness metrics like Equal parity ratio and statistical parity ratio.
The plot is provided by fairmodels
}
