% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_monitor.R
\name{plot_monitor}
\alias{plot_monitor}
\title{Plots monitor data}
\usage{
plot_monitor(
  STP = NULL,
  adversary_acc = NULL,
  adversary_losses = NULL,
  classifier_acc = NULL
)
}
\arguments{
\item{STP}{double, vector of Statistical Parity ratio value}

\item{adversary_acc}{double, vector containing adversarial models accuracy
for each epoch}

\item{adversary_losses}{double, vector of adversaries losses}

\item{classifier_acc}{double, vector of adversaries accuracy}
}
\value{
NULL - plots the visualizations
}
\description{
Plots visualizations of monitor metrics (STP ratio, adv accuracy, adv losses
and classifier losses) epoch by epoch. It is useful for the monitoring of
the learning process, thus the user can see if everything works properly.
To use this, the user has to set monitor on during the fair_train process.
}
\examples{

# presaved monitoring data
monitor2 <- torch_load(system.file("extdata","monitoring2",package="FairPAN"))

plot_monitor(monitor2$STP ,monitor2$adversary_acc, monitor2$classifier_acc,
             monitor2$adversary_losses)

}
