% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{preprocess}
\alias{preprocess}
\title{Preprocesses data for training}
\usage{
preprocess(
  data,
  target_name,
  sensitive_name,
  privileged,
  discriminated,
  drop_also,
  sample = 1,
  train_size = 0.7,
  test_size = 0.3,
  validation_size = 0,
  seed = NULL
)
}
\arguments{
\item{data}{list representing whole table of data.}

\item{target_name}{character, column name of the target variable. Selected column must be
interpretable as categorical.}

\item{sensitive_name}{character, column name of the sensitive variable.Selected column must
be interpretable as categorical.}

\item{privileged}{character meaning the name of privileged group}

\item{discriminated}{character meaning the name of discriminated group}

\item{drop_also}{character vector, column names of other columns to drop (like other
sensitive variables).}

\item{sample}{double from [0,1] setting size of our sample from original data set.
Default: 1}

\item{train_size}{double from [0,1] setting size of our train. Remember that
train_size+test_size+validation_size=1. Default=0.7}

\item{test_size}{double from [0,1] setting size of our test Remember that
train_size+test_size+validation_size=1. Default=0.3}

\item{validation_size}{double from [0,1] setting size of our validation. Remember that
train_size+test_size+validation_size=1. Default=0}

\item{seed}{sets seed for the sampling for code reproduction. Default=NULL}
}
\value{
list of prepared data (train_x,train_y,=sensitive_train,test_x,test_y,sensitive_test,
valid_x,valid_y,sensitive_valid,data_scaled_test,data_scaled_valid,data_test,
protected_test,data_valid,protected_valid)
}
\description{
This function prepapres provided dataset to be ready for the training process. It
makes data suitable for training functions, splits it into train, test and validation,
provides other data objects that are necessary for our training.
}
\examples{
data("adult")
preprocess(adult,"salary","sex",c("race"),sample=0.001,train_size=0.7,test_size=0.2,
validation_size=0.1,seed=7)
}
