% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretrain_net.R
\name{pretrain_net}
\alias{pretrain_net}
\title{Pretrains the neural network}
\usage{
pretrain_net(
  n_epochs = 15,
  model,
  dsl,
  model_type,
  learnig_rate,
  sensitive_test
)
}
\arguments{
\item{n_epochs}{indicates the number of epochs for neural network pretrain}

\item{model}{neural network model we want to train}

\item{dsl}{dataset_loader object for the training}

\item{model_type}{indicates which model (0 for adversarial, 1 for classifier)}

\item{learnig_rate}{float, setting learning rate for the training}

\item{sensitive_test}{test vector for sensitive variable used to calculate STP}
}
\value{

}
\description{
Pretrains the neural network
}
