% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_explainers.R
\name{Single_explainer}
\alias{Single_explainer}
\title{Provides fairmodels object for selected model/s}
\usage{
Single_explainer(target, model, data_set, protected, privileged)
}
\arguments{
\item{target}{numerical target of classification task}

\item{model}{first model we want to explain}

\item{data_set}{numerical table of predictors}

\item{protected}{numerical vector of sensitive variables}

\item{privileged}{string label of privileged class in protected}

\item{model2}{second model we want to explain}

\item{model3}{third model we want to explain}
}
\value{
fobject
}
\description{
The functions below provide fairmodels object for one, two or three neural network
models with the usage of DALEX explainer
}
