% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_explainers.R
\name{Triple_explainer}
\alias{Triple_explainer}
\title{Provides fairmodels object for selected three model}
\usage{
Triple_explainer(
  target,
  model,
  model2,
  model3,
  model_name,
  model_name2,
  model_name3,
  data_test,
  data_scaled_test,
  protected,
  privileged,
  batch_size,
  dev
)
}
\arguments{
\item{target}{numerical target of classification task}

\item{model}{first model we want to explain}

\item{model2}{second model we want to explain}

\item{model3}{third model we want to explain}

\item{data_test}{numerical table of predictors}

\item{data_scaled_test}{scaled matrix of numerical values representing predictors}

\item{protected}{numerical vector of sensitive variables}

\item{privileged}{string label of privileged class in protected}

\item{batch_size}{integer indicating a batch size used in dataloader.}

\item{dev}{device used to calculations (cpu or gpu)}
}
\value{
fobject - fairness object
}
\description{
The functions below provide fairmodels object for three neural network
models with the usage of DALEX explainer and fairmodels fairness_check.
}
