% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verbose_cat}
\alias{verbose_cat}
\title{Custom print function}
\usage{
verbose_cat(string, verbose = TRUE)
}
\arguments{
\item{string}{character (string) to print}

\item{verbose}{logical value indicating whether we want to print the string
or not}
}
\description{
Prints the provided string if verbose is TRUE.
}
