% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{calc_STP}
\alias{calc_STP}
\title{Calculates STP ratio}
\usage{
calc_STP(model, test_ds, sensitive, dev)
}
\arguments{
\item{model}{neural network model we want to evaluate}

\item{test_ds}{\code{dataset} object from torch used for making predictions
for STP ratio}

\item{sensitive}{numerical vector of sensitive variable}

\item{dev}{device used for calculations (cpu or gpu)}
}
\value{
float, STP ratio
}
\description{
Calculates Statistical Parity ((TP+FP)/(TP+FP+TN+FN)) ratio between
privileged and discriminated label for given model, which is the most
important fairness metric.
}
\examples{

\dontrun{
dev        <- if (torch::cuda_is_available())
                torch_device("cuda:0") else "cpu"

model     <- torch_load("./tests/zzz/clf2")
processed  <- torch_load("./tests/zzz/processed")

dsl        <- dataset_loader(processed$train_x, processed$train_y,
                             processed$test_x, processed$test_y,
                             batch_size = 5, dev = dev)

calc_STP(model, dsl$test_ds, processed$sensitive_test, dev)
}

}
