% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{dataset_loader}
\alias{dataset_loader}
\title{Creates datasets and dataloaders}
\usage{
dataset_loader(train_x, train_y, test_x, test_y, batch_size = 50, dev)
}
\arguments{
\item{train_x}{numeric, scaled matrix of predictors used for training}

\item{train_y}{numeric, scaled vector of target used for training}

\item{test_x}{integer, matrix of predictors used for testing}

\item{test_y}{integer, vector of predictors used for testing}

\item{batch_size}{integer indicating a batch size used in dataloader.
Default: 50}

\item{dev}{device used for calculations (cpu or gpu)}
}
\value{
list of two data sets and two dataloaders for train and test
respectively
}
\description{
Creates two torch datasets, from given train_x, test_x matrices and train_y,
test_y vectors and converts them into torch dataloaders with provided batch
size. It is used for both classifier and adversarial. During the initialization
NAs are ommited.
}
\examples{
train_x <- matrix(c(1,2,3,4,5,6),nrow=3)
train_y <- c(1,2,3)
test_x  <- matrix(c(1,2,3,4),nrow=2)
test_y  <- c(1,2)
dev     <- "cpu"
dataset_loader(train_x,train_y,test_x,test_y,batch_size=1,dev)
}
