% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{dataset_loader}
\alias{dataset_loader}
\title{Create data sets and dataloaders}
\usage{
dataset_loader(train_x, train_y, test_x, test_y, batch_size = 25)
}
\arguments{
\item{train_x}{numerical matrix with training predictor variables}

\item{train_y}{numerical vector with training target variables}

\item{test_x}{numerical matrix with test predictor variables}

\item{test_y}{numerical vector with test target variables}

\item{batch_size}{integer defining the size of batch size in dataloader}
}
\value{
list of two data sets and two dataloaders for train and test respectively
}
\description{
Create two data sets, from given ..x matrices and ..y vectors and converts them into
dataloaders with provided batch size.
}
