% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{eval_accuracy}
\alias{eval_accuracy}
\title{Evaluates model by calculating its accuracy}
\usage{
eval_accuracy(model, test_ds)
}
\arguments{
\item{model}{neural network model we want to evaluate}

\item{test_ds}{dataset object used for making test predictions and evaluation}
}
\value{
float accuracy of provided model
}
\description{
Evaluates model by calculating its accuracy
}
