% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{eval_accuracy}
\alias{eval_accuracy}
\title{Calculates accuracy}
\usage{
eval_accuracy(model, test_ds, dev)
}
\arguments{
\item{model}{net, nn_module, neural network model we want to evaluate}

\item{test_ds}{\code{dataset} object from torch used for making test
predictions and evaluation}

\item{dev}{device used for calculations (cpu or gpu)}
}
\value{
double, accuracy of provided model
}
\description{
This function evaluates the model by calculating its accuracy
}
\examples{

\dontrun{
dev        <- if (torch::cuda_is_available())
                 torch_device("cuda:0") else "cpu"
# presaved torch model
model      <- torch_load("./tests/zzz/clf2")
# presaved output of `preprocess` function
processed  <- torch_load("./tests/zzz/processed")
dsl        <- dataset_loader(processed$train_x, processed$train_y,
                             processed$test_x, processed$test_y,
                             batch_size = 5, dev = dev)

eval_accuracy(model, dsl$test_ds, dev)
}

}
