% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_pan.R
\name{explain_pan}
\alias{explain_pan}
\title{Creates explainer of PAN model}
\usage{
explain_pan(
  y,
  model,
  label,
  original_data,
  data,
  batch_size,
  dev,
  verbose = TRUE
)
}
\arguments{
\item{y}{numerical target of classification task}

\item{model}{net, nn_module, the model we want to explain}

\item{label}{character providing the label (name) to first model}

\item{original_data}{numerical list (table) of predictors}

\item{data}{scaled matrix of numerical values representing
predictors}

\item{batch_size}{integer indicating a batch size used in dataloader.}

\item{dev}{device used to calculations (cpu or gpu)}

\item{verbose}{logical indicating if we want to print monitored outputs or
not. Default: TRUE.}
}
\value{
DALEX model explainer
}
\description{
Creates DALEX explainer for PAN model (or other neural network)
All DALEX functions such as model_performance are possible to use on the
returned explainer.
}
\examples{

\dontrun{
dev <- "cpu"

adult <- fairmodels::adult

processed <- preprocess(
  adult,
  "salary",
  "sex",
  "Male",
  "Female",
  c("race"),
  sample = 0.8,
  train_size = 0.65,
  test_size = 0.35,
  validation_size = 0,
  seed = 7
)
# presaved torch model
model1 <- torch::torch_load(system.file("extdata","clf1",package="fairpan"))

dsl <- dataset_loader(processed$train_x, processed$train_y,
                      processed$test_x, processed$test_y,
                      batch_size=5, dev=dev)

explainer <- explain_pan(
  processed$test_y,
  model1,
  "classifier",
  processed$data_test,
  processed$data_scaled_test,
  batch_size = 5,
  dev = dev,
  verbose = FALSE
)
}


}
