% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_explainers.R
\name{explain_PAN}
\alias{explain_PAN}
\title{Provides fairmodels object for selected model}
\usage{
explain_PAN(
  target,
  model,
  model_name,
  data_test,
  data_scaled_test,
  batch_size,
  dev,
  verbose = TRUE
)
}
\arguments{
\item{target}{numerical target of classification task}

\item{model}{the model we want to explain}

\item{model_name}{character providing the label (name) to first model}

\item{data_test}{numerical list (table) of predictors}

\item{data_scaled_test}{scaled matrix of numerical values representing
predictors}

\item{batch_size}{integer indicating a batch size used in dataloader.}

\item{dev}{device used to calculations (cpu or gpu)}

\item{verbose}{logical indicating if we want to print monitored outputs or
not. Default: TRUE.}
}
\value{
fobject - fairness object
}
\description{
The functions below provide fairmodels object for one neural network
model with the usage of DALEX explainer and fairmodels fairness_check.
}
\examples{


}
