% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_train.R
\name{fair_train}
\alias{fair_train}
\title{Title}
\usage{
fair_train(
  n_ep_pan = 50,
  dsl,
  clf_model,
  adv_model,
  dev,
  sensitive_train,
  sensitive_test,
  batch_size = 50,
  learning_rate_adv = 0.001,
  learning_rate_clf = 0.001,
  lambda = 130
)
}
\arguments{
\item{lambda}{}
}
\value{

}
\description{
Title
}
