% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{make_preds}
\alias{make_preds}
\title{Makes binary predictions}
\usage{
make_preds(model, test_ds, dev)
}
\arguments{
\item{model}{neural network classification model}

\item{test_ds}{\code{dataset} object from torch used for making test
predictions}

\item{dev}{device used to calculations (cpu or gpu)}
}
\value{
integer (binary) vector of predictions
}
\description{
Makes binary predictions for classification model
}
\examples{

\dontrun{
dev       <- if (torch::cuda_is_available())
               torch_device("cuda:0") else "cpu"

model1    <- torch_load("./tests/zzz/preclf")
processed <- torch_load("./tests/zzz/processed")

dsl <- dataset_loader(processed$train_x, processed$train_y, processed$test_x,
                      processed$test_y, batch_size=5, dev=dev)

preds1   <- make_preds(model1,dsl$test_ds,dev)
}
}
