% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{make_preds_prob}
\alias{make_preds_prob}
\title{Makes probabilistic predictions}
\usage{
make_preds_prob(model, test_ds, dev)
}
\arguments{
\item{model}{neural network classification model}

\item{test_ds}{\code{dataset} object from torch used for making test
predictions}

\item{dev}{device used for calculations (cpu or gpu)}
}
\value{
float (probabilistic) vector of predictions
}
\description{
Makes probabilistic predictions for classification model
}
\examples{
dev       <- if (torch::cuda_is_available())
               torch_device("cuda:0") else "cpu"

model1    <- torch_load("./zzz/preclf")
processed <- torch_load("./zzz/processed")

dsl <- dataset_loader(processed$train_x, processed$train_y, processed$test_x,
                      processed$test_y, batch_size=5, dev=dev)

preds1   <- make_preds_prob(model1,dsl$test_ds,dev)

}
