% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{make_preds_prob}
\alias{make_preds_prob}
\title{Makes probabilistic predictions}
\usage{
make_preds_prob(model, test_ds, dev)
}
\arguments{
\item{model}{neural network classification model}

\item{test_ds}{\code{dataset} object from torch used for making test
predictions}

\item{dev}{device used for calculations (cpu or gpu)}
}
\value{
float (probabilistic) vector of predictions
}
\description{
Makes probabilistic predictions for neural network with classification task.
}
\examples{

dev       <-  "cpu"

# presaved torch model
model1    <- torch_load(system.file("extdata","preclf",package="fairpan"))

# presaved output of preprocess function
processed <- torch_load(system.file("extdata","processed",package="fairpan"))

dsl       <- dataset_loader(processed$train_x, processed$train_y,
             processed$test_x,processed$test_y, batch_size=5, dev=dev)

preds1    <- make_preds_prob(model1,dsl$test_ds,dev)


}
