% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_that.R
\name{plot_monitor}
\alias{plot_monitor}
\title{Plots visualizations that show the learning process}
\usage{
plot_monitor(
  STP = NULL,
  adversary_acc = NULL,
  adversary_losses = NULL,
  classifier_acc = NULL
)
}
\arguments{
\item{STP}{double vector of Statistical Parity ratio value}

\item{adversary_acc}{double vector of adversarials accuracy}

\item{adversary_losses}{double vector of adversaries losses}

\item{classifier_acc}{double vector of adversarials accuracy}
}
\value{
NULL - plots the visualizations
}
\description{
Plots visualizations that show the learning process
}
