% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretrain_net.R
\name{pretrain_net}
\alias{pretrain_net}
\title{Pretrains the neural network}
\usage{
pretrain_net(
  n_epochs = 15,
  model,
  dsl,
  model_type,
  learning_rate = 0.001,
  sensitive_test,
  dev
)
}
\arguments{
\item{n_epochs}{integer setting number of epochs for training. Default: 15}

\item{model}{neural network model we want to train}

\item{dsl}{dataset_loader object for the training}

\item{model_type}{indicates which model we train (0 for adversarial, 1 for classifier)}

\item{learning_rate}{float from [0,1] setting learning rate for model. Default: 0.001}

\item{sensitive_test}{test vector for sensitive variable used to calculate STP}

\item{dev}{device used to calculations (cpu or gpu)}
}
\description{
The function pretrain both the classifier and adversarial model. You can select which
model it is by setting model_type parameter
}
