% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_explainers.R
\name{Single_explainer}
\alias{Single_explainer}
\title{Provides fairmodels object for selected model}
\usage{
Single_explainer(
  target,
  model,
  model_name,
  data_set,
  data_scaled_test,
  test_y,
  protected,
  privileged,
  batch_size,
  dev
)
}
\arguments{
\item{target}{numerical target of classification task}

\item{model}{the model we want to explain}

\item{data_set}{numerical table of predictors}

\item{data_scaled_test}{scaled matrix of numerical values representing predictors}

\item{test_y}{integer, vector of predictors used for testing}

\item{protected}{numerical vector of sensitive variables}

\item{privileged}{string label of privileged class in protected}

\item{batch_size}{integer indicating a batch size used in dataloader.}

\item{dev}{device used to calculations (cpu or gpu)}
}
\value{
fobject - fairness object
}
\description{
The functions below provide fairmodels object for one neural network
model with the usage of DALEX explainer and fairmodels fairness_check.
}
