% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_PAN.R
\name{train_PAN}
\alias{train_PAN}
\title{Trains PAN model}
\usage{
train_PAN(
  n_ep_pan,
  dsl,
  clf_model,
  adv_model,
  dev,
  sensitive_train,
  sensitive_test,
  batch_size,
  learning_rate_adv,
  learning_rate_clf,
  lambda,
  verbose,
  monitor
)
}
\arguments{
\item{n_ep_pan}{number of epochs for PAN training}

\item{dsl}{dataset_loader object for classificator network}

\item{clf_model}{classifier model (preferably after pretrain)}

\item{adv_model}{adversarial model (preferably after pretrain)}

\item{dev}{device used to computation ("cuda" or "cpu")}

\item{sensitive_train}{integer vector of sensitive attribute used for training}

\item{sensitive_test}{integer vector of sensitive attribute used for testing}

\item{batch_size}{batch size used in adversarials dataset_loader}

\item{learning_rate_adv}{learning rate of adversarial}

\item{learning_rate_clf}{learning rate of classifier}

\item{lambda}{parameter regulating learning proccess (intuition: the bigger it is,
the fairer predictions).}
}
\description{
Trains Predictive Adversarial Network model, which means that it proceeds with the
mutual training of adverssarial on whole data set and classifier on a single mini batch.
The result is a fairer classifier.
}
